convert DonationForm, fix some uikit typings
