+++ bundles/uikit/icons/CheckboxChecked.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCheckSquare } from '@fortawesome/free-solid-svg-icons';
 
-const CheckboxOpen = props => {
+const CheckboxOpen = (props: any) => {
   return <FontAwesomeIcon {...props} icon={faCheckSquare} />;
 };
