+++ bundles/uikit/icons/CheckboxOpen.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSquare } from '@fortawesome/free-regular-svg-icons';
 
-const CheckboxOpen = props => {
+const CheckboxOpen = (props: any) => {
   return <FontAwesomeIcon {...props} icon={faSquare} />;
 };
